Create file GoalInput split goal input from app.js to goalinput.js Fix #18
